Deleting products from the client.

+++ controllers/admin.js
@@ -182,22 +182,22 @@ exports.postEditProduct = (req, res, next) => {
   }
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
+exports.deleteProduct = (req, res, next) => {
+  const { productId } = req.params;
   Product.findById(productId)
     .then(product => {
       if (!product) {
         return next(new Error("Product not found."));
       }
-      fileHelper.deleteFile(product.imageUrl);
-      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+      if (fileHelper.checkFile(product.imageUrl)) {
+        fileHelper.deleteFile(product.imageUrl);
+      }
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
     })
     .then(() => {
-      res.redirect("/admin/products");
+      res.status(200).json({ message: "Success!" });
     })
     .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
+      res.status(500).json({ message: "Deleting product failed!" });
     });
 };
